backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.2.0
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.114.1
h11==0.14.0
idna==3.8
passlib==1.7.4
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
PyJWT==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.9
ruff==0.6.5
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
